code/requirements.txt
django
psycopg2-binary
django-ninja
django-silk
django-ninja-simple-jwt
Pillow